Rename oauth backend as oauth2
